#2006: extract interface trackers from instance.go/manager.go into track.go
Group the flat pkg/vrrp package by concern. The interface-tracking
code (#1814) was split across instance.go (per-instance effective
priority) and manager.go (the singleton link watcher/poller), which
made the tracking subsystem hard to read as a unit and pushed
instance.go past 1200 lines. Collect it into a new track.go in the
SAME package — this is intra-package file reorganization, so there is
no import-path churn, no visibility change, and no cycle risk.

Moved (whole functions, byte-identical bodies):
  - from instance.go: getPriority, setTrackDown, trackedInterface
  - from manager.go:   ensureLinkWatcherLocked, runLinkWatcher,
    runLinkPoller, pollTrackedLinks, applyTrackedLinkState,
    seedTrackState, netlinkLinkState, linkAttrsUp

Behavior is preserved exactly: no change to the priority-cost demotion
clamp [1,254], the priority-0 resignation passthrough, the priority-255
owner exemption, the singleton-watcher latch under m.mu, the 1s poll
fallback, or any timing. The Manager struct's tracking fields stay in
manager.go (struct definition is single-file) and the vrrpInstance
trackDown field stays in instance.go for the same reason.

gofmt re-aligned a few struct-comment columns in the two touched files
(cosmetic; code tokens unchanged).

Validation: go build ./..., go test ./pkg/vrrp/... (ok), go vet
./pkg/vrrp/... (clean). The pre-existing track_test.go suite
(TestGetPriority_TrackMatrix, TestSetTrackDown_Transition,
TestLinkWatcher_*, TestPollTrackedLinks) passes unchanged.
